Drop rubocop fail level to W. Automated + manual fixes, manual review.

+++ app/controllers/challenges_controller.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 class ChallengesController < ApplicationController
   before_action :set_challenge, only: %i[show edit update destroy join entries]
   before_action :ensure_authenticated, only: %i[new create edit update destroy]
@@ -6,14 +8,14 @@ class ChallengesController < ApplicationController
   # GET /challenges
   # GET /challenges.json
   def index
-    @activeChallenges = Challenge.where('start_date <= ? AND (end_date > ? OR end_date IS NULL)', Date.current, Date.current).order("start_date ASC, end_date DESC")
-    @upcomingChallenges = Challenge.where('start_date > ?', Date.current).order("start_date DESC, end_date DESC")
-    @completedChallenges = Challenge.where('end_date <= ?', Date.current).order("start_date DESC, end_date DESC")
+    @activeChallenges = Challenge.where('start_date <= ? AND (end_date > ? OR end_date IS NULL)', Date.current, Date.current).order('start_date ASC, end_date DESC')
+    @upcomingChallenges = Challenge.where('start_date > ?', Date.current).order('start_date DESC, end_date DESC')
+    @completedChallenges = Challenge.where('end_date <= ?', Date.current).order('start_date DESC, end_date DESC')
   end
 
   # GET
   def entries
-    @challengeEntries = ChallengeEntry.includes(:submission).where(:challenge_id => @challenge.id).order("created_at DESC").paginate(:page => params[:page], :per_page => 25).includes(submission: :comments)
+    @challengeEntries = ChallengeEntry.includes(:submission).where(challenge_id: @challenge.id).order('created_at DESC').paginate(page: params[:page], per_page: 25).includes(submission: :comments)
   end
 
   # GET /challenge/1
@@ -22,27 +24,25 @@ class ChallengesController < ApplicationController
     # You can't join or leave a challenge after the start date.
     if Date.current < @challenge.start_date
       if params[:join]
-        Participation.find_or_create_by({ :user_id => current_user.id, :challenge_id => @challenge.id, :score => 0, :start_date => @challenge.start_date })
+        Participation.find_or_create_by({ user_id: current_user.id, challenge_id: @challenge.id, score: 0, start_date: @challenge.start_date })
       end
 
       if params[:leave]
-        participation = Participation.find_by({ :user_id => current_user.id, :challenge_id => @challenge.id })
-        if !participation.blank?
-          participation.destroy
-        end
+        participation = Participation.find_by({ user_id: current_user.id, challenge_id: @challenge.id })
+        participation.destroy if participation.present?
       end
     end
 
     if Date.current >= @challenge.start_date && (@challenge.end_date.blank? || Date.current < @challenge.end_date)
-      @currentParticipants = Participation.where({:challenge_id => @challenge.id, :active => true}).order("score DESC")
+      @currentParticipants = Participation.where({ challenge_id: @challenge.id, active: true }).order('score DESC')
     else
-      @currentParticipants = Participation.where({:challenge_id => @challenge.id}).order("score DESC, created_at ASC")
+      @currentParticipants = Participation.where({ challenge_id: @challenge.id }).order('score DESC, created_at ASC')
     end
     if @challenge.streak_based
       if @challenge.id == 1
-        @latestEliminations = Participation.where("challenge_id = 1 AND eliminated AND end_date <= :endDate AND end_date >= :startDate", {endDate: Date.current, startDate: (Date.current - 7.days)}).order("end_date DESC")
+        @latestEliminations = Participation.where('challenge_id = 1 AND eliminated AND end_date <= :endDate AND end_date >= :startDate', { endDate: Date.current, startDate: (Date.current - 7.days) }).order('end_date DESC')
       else
-        @allEliminations = Participation.where({:challenge_id => @challenge.id, :eliminated => true}).order("end_date DESC")
+        @allEliminations = Participation.where({ challenge_id: @challenge.id, eliminated: true }).order('end_date DESC')
       end
     end
   end
@@ -74,30 +74,22 @@ class ChallengesController < ApplicationController
     failure = false
 
     respond_to do |format|
-      if !@challenge.valid?
-        failure = true
-      end
-      if !@badge.valid?
-        failure = true
-      end
-      if !@badge_map.valid?
-        failure = true
-      end
+      failure = true unless @challenge.valid? && @badge.valid? && @badge_map.valid?
 
-      #specialized checks
+      # specialized checks
       if @challenge.start_date - Date.today < 7
-        @challenge.errors.add(:start_date, " should be at least a week out from today. Allow people to have sufficient advance notice to join!")
-        failure = true;
+        @challenge.errors.add(:start_date, ' should be at least a week out from today. Allow people to have sufficient advance notice to join!')
+        failure = true
       end
       if @challenge.streak_based && @challenge.postfrequency == 0
         @challenge.errors.add(:streak_based, " challenges cannot have a post frequency of 'None'.")
-        failure = true;
+        failure = true
       end
-      if @challenge.postfrequency != 0 && !@badge_map.required_score.blank?
+      if @challenge.postfrequency != 0 && @badge_map.required_score.present?
         possibleScore = ((@challenge.end_date - @challenge.start_date) / @challenge.postfrequency).to_i
         if @badge_map.required_score > possibleScore
           @badge_map.errors.add(:required_score, " is impossible to achieve within the dates and post frequency specified (only #{possibleScore} submissions possible).")
-          failure = true;
+          failure = true
         end
       end
 
@@ -113,10 +105,12 @@ class ChallengesController < ApplicationController
 
         User.where.not(id: current_user.id).each do |u|
           # If the user has submitted within the last two weeks, send a notification of a starting challenge.
-          next if Submission.find_by("created_at >= ? and user_id = ?", Date.today - 14.day, u.id).nil?
+          if Submission.find_by('created_at >= ? and user_id = ?', Date.today - 14.days, u.id).nil?
+            next
+          end
 
           Notification.create(body: "#{@challenge.name} has been created by #{current_user.username}. Check it out, and consider signing up!",
-                              source_type: "Challenge",
+                              source_type: 'Challenge',
                               source_id: @challenge.id,
                               user_id: u.id,
                               url: "/challenges/#{@challenge.id}")
@@ -127,8 +121,7 @@ class ChallengesController < ApplicationController
     end
   end
 
-  def edit
-  end
+  def edit; end
 
   def update
     failure = false
@@ -137,20 +130,20 @@ class ChallengesController < ApplicationController
       newChallenge = Challenge.new(allowed_challenge_params)
       newBadgeMap = BadgeMap.new(allowed_badge_map_params)
 
-      #specialized checks
+      # specialized checks
       if Date.current < @challenge.start_date && newChallenge.start_date - Date.today < 7
-        @challenge.errors.add(:start_date, " should be at least a week out from today. Allow people to have sufficient advance notice to join!")
-        failure = true;
+        @challenge.errors.add(:start_date, ' should be at least a week out from today. Allow people to have sufficient advance notice to join!')
+        failure = true
       end
       if newChallenge.streak_based && newChallenge.postfrequency == 0
         @challenge.errors.add(:streak_based, " challenges cannot have a post frequency of 'None'.")
-        failure = true;
+        failure = true
       end
-      if newChallenge.postfrequency != 0 && !newBadgeMap.required_score.blank?
+      if newChallenge.postfrequency != 0 && newBadgeMap.required_score.present?
         possibleScore = ((newChallenge.end_date - newChallenge.start_date) / newChallenge.postfrequency).to_i
         if newBadgeMap.required_score > possibleScore
           @badge_map.errors.add(:required_score, " is impossible to achieve within the dates and post frequency specified (only #{possibleScore} submissions possible).")
-          failure = true;
+          failure = true
         end
       end
 
@@ -179,24 +172,24 @@ class ChallengesController < ApplicationController
   end
 
   private
-    # Use callbacks to share common setup or constraints between actions.
-    def set_challenge
-      @challenge = Challenge.find(params[:id])
-      @badge_map = BadgeMap.find_by(challenge_id: @challenge.id)
-      @badge_maps = BadgeMap.where(challenge_id: @challenge.id).order(:prestige)
-      @badge = Badge.find(@badge_map.badge_id)
-    end
 
-    def allowed_badge_map_params
-      params.require(:badge_map).permit(:required_score, :prestige, :description)
-    end
+  # Use callbacks to share common setup or constraints between actions.
+  def set_challenge
+    @challenge = Challenge.find(params[:id])
+    @badge_map = BadgeMap.find_by(challenge_id: @challenge.id)
+    @badge_maps = BadgeMap.where(challenge_id: @challenge.id).order(:prestige)
+    @badge = Badge.find(@badge_map.badge_id)
+  end
 
-    def allowed_challenge_params
-      params.require(:challenge).permit(:name, :description, :nsfw_level, :start_date, :end_date, :streak_based, :postfrequency)
-    end
+  def allowed_badge_map_params
+    params.require(:badge_map).permit(:required_score, :prestige, :description)
+  end
 
-    def allowed_badge_params
-      params.require(:badge).permit(:name, :avatar)
-    end
+  def allowed_challenge_params
+    params.require(:challenge).permit(:name, :description, :nsfw_level, :start_date, :end_date, :streak_based, :postfrequency)
+  end
 
+  def allowed_badge_params
+    params.require(:badge).permit(:name, :avatar)
+  end
 end
